Fixes "/list" message if only one player online
Right now if there is only one player online, it still says "There are 1 players online!" which does not make grammatical sense, so this commit changes it to "There is 1 player online!". To do this I created a new int variable called playerAmount, so I would not have to call players.size() three times.